refactor: Enum으로 Spot과 NonSpot 구분
