+++ backend/src/main/java/hanglog/trip/domain/item/Item.java
@@ -1,11 +1,15 @@
-package hanglog.trip.domain;
+package hanglog.trip.domain.item;
 
+import static jakarta.persistence.EnumType.STRING;
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
+import hanglog.trip.domain.daylog.DayLog;
+import hanglog.trip.domain.item.type.ItemType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
@@ -23,6 +27,10 @@ public class Item {
     private Long id;
 
     @Column(nullable = false)
+    @Enumerated(value = STRING)
+    private ItemType itemType;
+
+    @Column(nullable = false)
     private String title;
 
     @Column(nullable = false)
@@ -34,7 +42,13 @@ public class Item {
     @JoinColumn(name = "day_log_id", nullable = false)
     private DayLog dayLog;
 
-    public Item(final String title, final Long rating, final String memo, final DayLog dayLog) {
+    public Item(final ItemType itemType,
+                final String title,
+                final Long rating,
+                final String memo,
+                final DayLog dayLog
+    ) {
+        this.itemType = itemType;
         this.title = title;
         this.rating = rating;
         this.memo = memo;
