+++ backend/src/main/java/hanglog/trip/domain/item/type/ItemType.java
@@ -1,0 +1,6 @@
+package hanglog.trip.domain.item.type;
+
+public enum ItemType {
+    SPOT,
+    NON_SPOT
+}
